Reform TypeScript API
The purpose of api.ts is for people to be able to (type-)safely
construct API objects; but, with the previous formulation in terms of
Partial<T> types, it's possible to end up with incomplete objects.

By using the types themselves as arguments, the required fields are
required in the `fields` argument.